Déplace le composant `FieldMessage`dans son propre fichier et change la couleur du texte par défaut
