+++ packages/ui/src/design-system/Field/Field.tsx
@@ -1,5 +1,5 @@
 import classNames from 'classnames';
-import {Icon} from '@design-system/Icon';
+import {FieldMessage, FieldMessageProps} from './FieldMessage';
 
 export type FieldState =
   | 'default'
@@ -8,24 +8,6 @@ export type FieldState =
   | 'success'
   | 'warning'
   | 'disabled';
-
-export const stateToTextColor: Record<FieldState, string> = {
-  info: 'text-info-1',
-  error: 'text-error-1',
-  success: 'text-success-1',
-  warning: 'text-warning-1',
-  default: 'text-grey-8',
-  disabled: 'text-grey-6',
-};
-
-const stateToIcon: Record<FieldState, string> = {
-  info: 'information-fill',
-  error: 'spam-fill',
-  success: 'checkbox-circle-fill',
-  warning: 'alert-fill',
-  default: '',
-  disabled: '',
-};
 
 type Props = {
   /** Titre */
@@ -79,40 +61,5 @@ export const Field = ({
       {children}
       <FieldMessage state={state} message={message} />
     </div>
-  );
-};
-
-export type FieldMessageProps = {
-  /** État */
-  state?: FieldState;
-  /** Message d'état affiché en dessous du champ */
-  message?: string;
-  /** Pour surcharger les styles du container du message */
-  messageClassName?: string;
-};
-
-/**
- * Affiche un message additionnel au-dessous d'un champ
- */
-export const FieldMessage = ({
-  message,
-  state = 'default',
-  messageClassName,
-}: FieldMessageProps) => {
-  return (
-    message !== undefined && (
-      <div
-        className={classNames(
-          'flex items-center',
-          stateToTextColor[state],
-          messageClassName
-        )}
-      >
-        {state !== 'disabled' && state !== 'default' && (
-          <Icon icon={stateToIcon[state]} size="sm" className="mr-1" />
-        )}
-        <span className="text-xs">{message}</span>
-      </div>
-    )
   );
 };
