+++ packages/ui/src/design-system/Field/index.ts
@@ -1,1 +1,2 @@
 export * from './Field';
+export * from './FieldMessage';
